Project restructuralized
