--- ConsoleKaraoke/Karaoke.Abstractions/TableRow.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Karaoke.Abstractions
-{
-    public class TableRow
-    {
-        public string Syllable { get; set; }
-        public string Tone { get; set; }
-        public string Length { get; set; }
-        public int Pause { get; set; }
-    }
-}
