+++ ConsoleKaraoke/Karaoke.Songs/SongListCreator.cs
@@ -1,8 +1,8 @@
 ﻿using Karaoke.Abstractions;
 using System;
+using System.Buffers.Text;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -11,18 +11,18 @@ namespace Karaoke.Songs
     public static class SongListCreator
     {
         // A method that creates the List of songs based on all classes that inherit from the SongBase class
-        public static List<SongBase> CreateSongsList()
+        public static List<ISong> CreateSongsList()
         {
-            var songsList = new List<SongBase>();
-            var assemblies = AppDomain.CurrentDomain.GetAssemblies(); // Since the base class SongBase is a in different assembly than Song1 and Song2
+            var songsList = new List<ISong>();
+            var assemblies = AppDomain.CurrentDomain.GetAssemblies();
 
             var songTypes = assemblies.SelectMany(assembly => assembly.GetTypes())
-                                      .Where(t => t.IsSubclassOf(typeof(SongBase)))
+                                      .Where(s => typeof(ISong).IsAssignableFrom(s) && !s.IsInterface && !s.IsAbstract)
                                       .ToList();
 
             foreach (var type in songTypes)
             {
-                if (Activator.CreateInstance(type) is SongBase song)
+                if (Activator.CreateInstance(type) is ISong song)
                 {
                     songsList.Add(song);
                 }
