+++ ConsoleKaraoke/Karaoke.Songs/TableRow.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Karaoke.Songs
+{
+    public class TableRow
+    {
+        public string Syllable { get; set; }
+        public string Tone { get; set; }
+        public string Length { get; set; }
+        public int Pause { get; set; }
+    }
+}
